Move cell-method parsing to own sourcefile, as it is not part of netcdf.saver.
